+++ tests/utils/test_db.py
@@ -34,7 +34,6 @@ from alembic.script import ScriptDirectory
 from sqlalchemy import MetaData, Table
 from sqlalchemy.sql import Select
 
-from airflow.exceptions import AirflowException
 from airflow.models import Base as airflow_base
 from airflow.settings import engine
 from airflow.utils.db import (
@@ -49,7 +48,6 @@ from airflow.utils.db import (
     # guard against removing it from utils.db accidentally
     create_session,  # noqa: F401
     downgrade,
-    initdb,
     resetdb,
     upgradedb,
 )
@@ -320,57 +318,3 @@ class TestDb:
             mock_session, task_fail_table, mock_select, dangling_task_fail_table_name
         )
         mock_session.rollback.assert_called_once()
-
-
-class TestRunDBManager:
-    def test_fab_db_manager_is_default(self):
-        from airflow.providers.fab.auth_manager.models.db import FABDBManager
-
-        run_db_manager = RunDBManager()
-        assert run_db_manager._managers == [FABDBManager]
-
-    def test_defining_table_same_name_as_airflow_table_name_raises(self):
-        from sqlalchemy import Column, Integer, String
-
-        run_db_manager = RunDBManager()
-        manager = run_db_manager._managers[0]
-        # Add dag_run table to metadata
-        mytable = Table(
-            "dag_run", manager.metadata, Column("id", Integer, primary_key=True), Column("name", String(50))
-        )
-        manager.metadata._add_table("dag_run", None, mytable)
-        with pytest.raises(AirflowException, match="Table 'dag_run' already exists in the Airflow metadata"):
-            run_db_manager.validate()
-
-    @mock.patch("airflow.utils.db.RunDBManager")
-    def test_init_db_calls_rundbmanager(self, mock_rundbmanager, session):
-        initdb(session=session)
-        mock_rundbmanager.return_value.initdb.assert_called_once_with(session)
-        mock_rundbmanager.return_value.upgradedb.assert_not_called()
-        mock_rundbmanager.return_value.downgrade.assert_not_called()
-
-    @mock.patch("airflow.utils.db.RunDBManager")
-    @mock.patch("alembic.command")
-    def test_upgradedb_or_downgrade_dont_call_rundbmanager(
-        self, mock_alembic_command, mock_rundbmanager, session
-    ):
-        upgradedb(session=session)
-        mock_alembic_command.upgrade.assert_called_once_with(mock.ANY, revision="heads")
-        downgrade(to_revision="base")
-        mock_alembic_command.downgrade.assert_called_once_with(mock.ANY, revision="base", sql=False)
-        mock_rundbmanager.return_value.initdb.assert_not_called()
-        mock_rundbmanager.return_value.upgradedb.assert_not_called()
-        mock_rundbmanager.return_value.downgrade.assert_not_called()
-
-    @mock.patch("airflow.providers.fab.auth_manager.models.db.FABDBManager")
-    def test_rundbmanager_calls_dbmanager_methods(self, mock_fabdb_manager, session):
-        ext_db = RunDBManager()
-        # initdb
-        ext_db.initdb(session=session)
-        mock_fabdb_manager.return_value.initdb.assert_called_once()
-        # upgradedb
-        ext_db.upgradedb(session=session)
-        mock_fabdb_manager.return_value.upgradedb.assert_called_once()
-        # downgradedb
-        ext_db.downgradedb(session=session)
-        mock_fabdb_manager.return_value.downgradedb.assert_called_once()
